refactor: move Settings.js to Settings/index.js
