+++ modules/Settings/index.js
@@ -3,19 +3,19 @@ import { useTranslation } from 'react-i18next'
 import PropTypes from 'prop-types'
 import { Link, Route, Switch, useRouteMatch } from 'react-router-dom'
 
-import Profile from './settings/Profile'
-import Organization from './settings/Organization'
-import Organizations from './settings/Organizations'
-import OrganizationNew from './settings/OrganizationNew'
-import Tags from './settings/Tags'
-import Tag from './settings/Tag'
-import SettingsCSProfile from './settings/CustomerServiceProfile'
-import Members from './settings/Members'
-import FAQs from './settings/FAQs'
-import FAQ from './settings/FAQ'
-import Categories from './settings/Categories'
-import Category from './settings/Category'
-import CategorySort from './settings/CategorySort'
+import Profile from './Profile'
+import Organization from './Organization'
+import Organizations from './Organizations'
+import OrganizationNew from './OrganizationNew'
+import Tags from './Tags'
+import Tag from './Tag'
+import SettingsCSProfile from './CustomerServiceProfile'
+import Members from './Members'
+import FAQs from './FAQs'
+import FAQ from './FAQ'
+import Categories from './Categories'
+import Category from './Category'
+import CategorySort from './CategorySort'
 
 export default function Settings(props) {
   const { path } = useRouteMatch()
